Cleaned out index js a little more and added a new setup file for ui elements.
